Add interactive and timing options
When optional interactive option is on, the user is prompted to press Enter to
move forward in the execution of script. When he is not using the interactive
mode, the timing option is used to wait after plots have been made.

Prefer the non-interactive mode for a quick test of the scripts. Prefer the
interactive mode when reading code along execution, to understand tunacell
commands.